+++ src/overviewframe.h
@@ -4,8 +4,13 @@
 #pragma once
 
 #include <QFrame>
+#include <QFile>
+#include <QTextStream>
 
 class QAbstractItemModel;
+class QProgressDialog;
+class QFile;
+class QTextStream;
 
 namespace Ui {
 class OverviewFrame;
@@ -16,6 +21,7 @@ namespace WalletGUI {
 class WalletModel;
 class MiningManager;
 class CopiedToolTip;
+class CSVTransactionsExporter;
 
 class OverviewFrame : public QFrame
 {
@@ -27,10 +33,13 @@ public:
     ~OverviewFrame();
 
     void setMainWindow(QWidget* mainWindow);
-    void setTransactionsModel(QAbstractItemModel* model);
+    void setTransactionsModel(WalletModel* transactionsModel);
     void setWalletModel(WalletModel* walletModel);
     void setMiningManager(MiningManager* miningManager);
     void setMinerModel(QAbstractItemModel* model);
+
+public slots:
+    void exportToCSV();
 
 signals:
     void copiedToClipboardSignal();
@@ -39,10 +48,36 @@ signals:
 private:
     QScopedPointer<Ui::OverviewFrame> m_ui;
     QWidget* m_mainWindow;
-    QAbstractItemModel* m_transactionsModel;
+    WalletModel* m_transactionsModel;
+    CSVTransactionsExporter* m_csvExporter;
 
     void rowsInserted(const QModelIndex& parent, int first, int last);
     bool eventFilter(QObject* object, QEvent* event) override;
+};
+
+class CSVTransactionsExporter : public QObject
+{
+    Q_OBJECT
+
+public:
+    CSVTransactionsExporter(WalletModel* transactionsModel, QWidget* parent);
+
+    void start();
+    void cancel();
+
+private:
+    QWidget* parent_;
+    QFile file_;
+    QTextStream csv_;
+    QProgressDialog* exportProgressDlg_;
+    WalletModel* transactionsModel_;
+
+    void finish();
+    void fetched();
+    void writeHeader();
+    void writeData();
+    void close();
+    void reset();
 };
 
 }